Basic - Version Information Modification
Simple for starters, changing version information to indicate a new experimental version in development; what I plan to do here is minimize, shrink, and optimize the code making it less complex, more simple to handle and manage, more dumbed down and straightforward. And eventually afterward, once this version is perfected and finalized, I will attempt a new GUI version of ths program/script based on/off this code since it'd just be easier to do so compared to now. Doing it now means I'd have to create the GUI version dependant upon the current version, create this version, then rewrite/update the GUI version to match.